+++ src/main.rs
@@ -1,4 +1,4 @@
-const PROGRAM_VERSION: &str = "0.14.0";
+const PROGRAM_VERSION: &str = "0.16.0";
 
 fn main() -> std::process::ExitCode {
     b9::cli::run(PROGRAM_VERSION)